Adding return values for WM_NCHITTEST.
One problem I cannot make them type-safe as the other constants, because
WndProc always must return LResult type. The solution is to create a
function to convert HitTestResult to LRESULT. This is not the only return
type I'm going to have, there are others. It would be awkward to have a
conversion function for each return type, so I define a class ReturnValue
with a single function toLResult :: v -> LRESULT. Every return type must
implement this class.